Move unit tests into their own folder, ensure they pass!
