+++ tests/unit/fake_snapd.py
