+++ tox.ini
@@ -12,10 +12,10 @@ lib_path = {toxinidir}/lib/charms/operator_libs_linux/
 all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
 
 [testenv]
-basepython = python3
 setenv =
   PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
   PYTHONBREAKPOINT=ipdb.set_trace
+  PY_COLORS=1
 passenv =
   PYTHONPATH
   HOME
@@ -60,5 +60,6 @@ deps =
     -r{toxinidir}/requirements.txt
     pyfakefs
 commands =
-	coverage run --source={[vars]lib_path} -m pytest -v --tb native {posargs}
+	coverage run --source={[vars]lib_path} \
+        -m pytest --ignore={[vars]tst_path}integration -v --tb native {posargs}
     coverage report
